==> sentiment140/00_Determining_Next_Steps.ipynb <==
## Determining Next Steps

TODO:

- Processed folder has great nomenclature for X but poor for y (and missing): Fix

Current POC

- POC9: Extra Trees and VarImp plots

Next POC

- adaboost and gradient boosting
- SVC
- ensembles

Other Considerations

- Try TextBlob?

Better Representation

- TF-IDF (see DSB)
- Bag of N-grams
- LSA/SVD

Feature Engineering

- TextLength, etc
- CosineSimilarity 
- count of: punctuations, ascii chars, USERNAMEs, EMOJIs, URLs
- tweet starts with USERNAME, or EMOJI, or URL, etc.
- count of swear words, or negative words, or positive words, etc. (need lists)
- semantic analysis research

Feature Selection

- Random Forest + VarImp plot 
- LASSO

Direction

- define goal and purpose.. what's the use case? App idea?
- POC to full-blown: do not use full dataset unless it is warranted (learning curves)
- multi-processing cleanup effort: verify whether developing a reproducible async method for cleanup is warranted; might be worth if iterating on representations and also for new twitterbot data
- define evaluation metric better? If the goal is overall performance, the MCC (from sklearn.metrics import matthews_corrcoef) is probably preferred as it takes into account the entirety of the confusion matrix

---
