==> sentiment140/00_POC15_Assessment_Predictions.ipynb <==
# Twitter Sentiment Analysis - POC
---


import re
import os
import time
import json

import numpy as np
import pandas as pd
import scipy.sparse as sp

import cleanup_module_POC as Cmod

from textblob import TextBlob
from sklearn.pipeline import Pipeline
from sklearn.naive_bayes import MultinomialNB
from sklearn.linear_model import LogisticRegression
from sklearn.model_selection import cross_val_score
from sklearn.model_selection import train_test_split

# time notebook
start_notebook = time.time()

# load minimally prepared X, y train subsets
raw_path = os.path.join("..","data","1_raw","sentiment140")
X_train = pd.read_csv(os.path.join(raw_path, "X_train.csv"))
y_train = pd.read_csv(os.path.join(raw_path, "y_train.csv"))

### Sample down to $m\approx120k$

X, X_rest, y, y_rest = train_test_split(X_train, y_train, test_size=0.9, random_state=42)

# check
print(f'Dataset size: {len(X):0.0f}')
print(f'Target distribution: {sum(y["target"])/len(y):0.3f}')

# transform into arrays
X_array = np.array(X.iloc[:, 2]).ravel()
y_array = y.iloc[:,0].ravel()

preprocess_pipeline = Pipeline([
    ("document_to_wordcount", Cmod.DocumentToWordCounterTransformer()),
    ("wordcount_to_vector", Cmod.WordCounterToVectorTransformer(vocabulary_size=30000)),
])

 proc_dir = os.path.join("..","data","3_processed","sentiment140")
try:
    # load X_train_transformed 
    X_train_transformed = sp.load_npz(os.path.join(proc_dir, "X_train_transformed_BoW_120k_30k.npz"))
    print('Loading X_train_transformed...')
except:
    print('Preprocessing X_array...')
    start_pipeline = time.time()
    X_train_transformed = preprocess_pipeline.fit_transform(X_array) 
    # save preprocessed data
    sp.save_npz(os.path.join(proc_dir, 'X_train_transformed_BoW_120k_30k.npz'), X_train_transformed)
    # print processing time
    mins, secs = divmod(time.time() - start_pipeline, 60)
    print(f'Preprocessing time: {mins:0.0f} minute(s) and {secs:0.0f} second(s).')
    # Preprocessing time: 2 minute(s) and 41 second(s).

X_train_transformed

def make_binary(val):
    if val <= 0:
        return 0
    else:
        return 1
        
def extract_polarity(array):
    bin_vals = []
    for entry in array:
        blob = TextBlob(entry)
        polarity = blob.sentiment.polarity
        bin_vals.append(make_binary(polarity))
    return bin_vals

y_preds = extract_polarity(X_array)

from sklearn.metrics import accuracy_score
print(f'TexBlob\'s accuracy: {accuracy_score(y_array, y_preds):0.4f}')

def calc_accuracy(y_true, y_pred):
    sum_ones = []
    for yt, yp in zip(y_array, y_pred):
        if yt == yp:
            sum_ones.append(1)
        else:
            sum_ones.append(0)
    return sum(sum_ones)/len(sum_ones)

print(f'TexBlob\'s accuracy: {calc_accuracy(y_array, y_preds):0.4f}') 

But that's without any stemming, nothing.

### Cross Validation 

NB_clf = MultinomialNB()
log_clf = LogisticRegression(solver="liblinear", random_state=42)

score = cross_val_score(NB_clf, X_train_transformed, y_array, cv=5, verbose=0, scoring='accuracy')
print(f'Accuracy: {round(score.mean(),4):0.4f} (+/- {np.std(score):0.4f})')

score = cross_val_score(log_clf, X_train_transformed, y_array, cv=5, verbose=0, scoring='accuracy')
print(f'Accuracy: {round(score.mean(),4):0.4f} (+/- {np.std(score):0.4f})')

### Peek at some predictions

new_array1 = np.array([
                      "loving this feeling amazing happy",
                      "yay we're going bowling - so excited",
                      "what a wonderful day!",
                       "they can go to hell, idiots",
                      "I hate you and I hate everyone",
                      "what a poor excuse - sorry but they're terrible liars"
                      ])

new_array2 = np.array([
                       "they can go to hell, idiots",
                      "I hate you and I hate everyone",
                      "what a poor excuse - sorry but they're terrible liars",
                      "loving this feeling amazing happy",
                      "yay we're going bowling - so excited",
                      "what a wonderful day!"
                      ])

narray1_transformed = preprocess_pipeline.fit_transform(new_array1)
narray2_transformed = preprocess_pipeline.fit_transform(new_array2)

narray1_transformed

log_clf = LogisticRegression(solver="liblinear", random_state=42)
log_clf.fit(X_train_transformed, y_array)

log_clf.predict_proba(narray1_transformed)

log_clf.predict(narray1_transformed)

log_clf.predict_proba(narray2_transformed)

log_clf.predict(narray2_transformed)

new_array1

for string in new_array1:
    blob = TextBlob(string)
    print(string + f' {blob.sentiment.polarity:0.3f}')

for string in new_array2:
    blob = TextBlob(string)
    print(string + f' {blob.sentiment.polarity:0.3f}')

---
