==> sentiment140/00_POC1_Twitter_Sentiment_Analysis.ipynb <==
# Twitter Sentiment Analysis - POC


## Project Structure

This is a proof of concept project for the main project. I'm first running the entire project structure with a very small dataset, avoiding all the complexity of using a large dataset, multiprocessing, and so forth, to focus on the project iself.

The project structure is:

1. set expectations and define the problem, question the project is addressing, or product being developed 
2. preliminary exploratory data analysis (EDA), just enough to figure out whether some pre-dataset-splitting cleanup needs to be performed
3. split dataset into trainining and test subsets; set the test subset aside and do not peek further to avoid data leakage
4. create a cleanup pipeline for the training data that can be re-applied to the test data
5. create an ML pre-processing pipeline that can be re-applied as well
6. using cross-validation, evaluate a variety models without hyperparameter tuning to establish some baselines
7. short-list promising models for further hyperparemeter tuning
8. iterate on any phase of the project as needed (refine the question/problem, recreate the cleanup/pre-processing)
9. consider feature selection and feature engineering
10. decide on a final cleanup and processing pipeline
11. settle on a final model
12. re-apply all cleanup and processing steps to the test set
13. evaluate the final model using the test set once and estimate the uncertainty around this generalization
14. create a presentation including how this model would be deployed, evaluated, and maintained in production

---

## 1. Project Definition

The project's goal is to simply to develop a machine learning model that successfully predicts or classifies whether a Tweet is positive or negative, and possibly how strongly so, as a way to automatically generate reports as an input to an app.

The app is beyond the scope of the project - this could be any app that uses short text input from users and needs to, as part of its process, evaluate the "sentiment" of the text. An example would be Twitter users who download their Tweets for the last year and want to know how positive or negative their posts were, in a timeline for example, or as aggregate in comparison with a cohort of friends. Another example would be an app that tracks positive/negative sentiments around a topic, by using hashtags and whatnot, in conjunction with this sentiment evaluation.

The scope of this project is to provide a model that most accurately predicts the sentiment of a Tweet - whether negative or positive, and does so within a reasonable amount of time, but not as fast as possible (not real time). The evaluation criteria of what constitutes "success" and "most accurate prediction" is TBD. I might use ROC/AUC curves and whatnot. Accuracy alone isn't enough without considering recall.

---

## 2. Preliminary EDA

For the POC, I'm using a very small dataset - but first we examine the entire dataset.

import os
import time
import numpy as np
import pandas as pd 

# time notebook
start_notebook = time.time()

# load raw dataset
raw_path = os.path.join("..","data","1_raw","sentiment140")
filename = "training.1600000.processed.noemoticon.csv"
filepath = os.path.join(raw_path, filename)

raw = pd.read_csv(filepath, encoding='latin-1', usecols=[0,1,4,5])
raw.columns = ["target","ID","username","tweet"]

There are four columns:
- target: is the target of prediction, 0 means negative, 4 means positive
- ID: is the Tweet ID for each Tweet
- username: is the username of the Tweeter
- tweet: is the text of the Tweet

### Preview data

raw.head()

raw.tail()

### Target distribution

There are 1,599,999 rows and the target is evenly distributed:

raw[["target","ID"]].groupby("target").count()

### Missing values

There are no NA values:

raw.isnull().sum()

### Deduplication

There are 3370 Tweets which are suspect. Half of them are the duplicated values, the other half are involved in this and look suspicious as they show with both positive and negative targets:

all_dupes = raw[raw.duplicated(subset=['ID'], keep=False)].sort_values(by=['ID'])
all_dupes.head(6)

len(all_dupes)

Verifying they are indeed all positive/negative "Twin Tweets":

all_dupes_grouped = all_dupes[["ID","target"]].groupby("ID").count()
all_dupes_grouped.head(3)

all_dupes_grouped[all_dupes_grouped["target"] != 2]

Since these tweets are not useful, I'm dropping them from the entire dataset. Deduplication is a complex procedure that I will not be reproducing in a "cleanup pipeline", my test set won't have duplicates.

raw_deduped = raw.drop_duplicates(subset="ID", keep=False)

# double check
assert len(raw) - len(raw_deduped) == len(all_dupes)

We expect the distribution of the target to be equally balanced as before:

raw_deduped[["target","ID"]].groupby("target").count()

We got rid of ID dupes, and we know there are username dupes (we don't care about that), but what about retweets? There might be tons of Tweet dupes, so we need to check for that - it appears as though our dataset creators made sure to avoid retweets!

text_dupes = raw_deduped[raw_deduped.duplicated(subset=['ID'], keep=False)].sort_values(by=['tweet'])
text_dupes.head(6)

# reindex
raw_deduped.index = range(len(raw_deduped))

raw_deduped.tail()

### Split X, y

Here I just split my data into a feature matrix X and its target vector y:

X = raw_deduped.iloc[:, 1:4]
y = raw_deduped.iloc[:, 0]

### Recode target

A final and minimal cleanup is to recode the target values to 0s and 1s. While in sentiment analysis the polarity of a sentiment is typically a continuum from -1 to +1, since we don't have a grading this coding makes no sense. I'll frame this discretized binary target as a vanilla `IsPositive` feature:

y_df = pd.DataFrame({'target':y}) 
y_df[y_df["target"]==4] = 1
y_df.tail()

---

## 3. Split into Trainining & Test Sets

import numpy as np
from sklearn.model_selection import train_test_split

X_train, X_test, y_train, y_test = train_test_split(X, y_df, test_size=0.25, random_state=42)

### Set Test set aside

Save training and test sets, and indices as well (although arguably the procedures is reproducible by setting the same random seed above).

X_train.head()

y_train.head()

# save train and test indices
trainpath = os.path.join(raw_path, "train_ix.npy")
testpath = os.path.join(raw_path, "test_ix.npy")

with open(trainpath, 'wb') as f:
    np.save(f, np.array(y_train.index))
    
with open(testpath, 'wb') as f:
    np.save(f, np.array(y_test.index))

# Save train and test sets
# Note: index=True does not preserve split indices
X_train.to_csv(os.path.join(raw_path, "X_train.csv"), index=False) 
X_test.to_csv(os.path.join(raw_path, "X_test.csv"), index=False)
y_train.to_csv(os.path.join(raw_path, "y_train.csv"), index=False)
y_test.to_csv(os.path.join(raw_path, "y_test.csv"), index=False)

---

mins, secs = divmod(time.time() - start_notebook, 60)
print(f'Total running time: {mins:0.0f} minute(s) and {secs:0.0f} second(s)')
