==> sentiment140/00_TEST_determine_next_steps.ipynb <==
## Determining Next Steps

Current State

- NB and LR at 75%, unigram BoW, m,n=12k,1k 
- plot learning curves for a given m,n size, keeping n constant, or m constant, and either NB or LR (has to be fast)

Extra:

- Look at actual predictions (text)? 
- Are we predicting positive/negative Tweets or positive/negative emoticons? 
- If we apply to Tweets without emoticons, predict sentiment correctly? (would need human-made labels)
- Use TextBlob? Another method (more research)?
- App idea


Better Representation

- TF-IDF
- Ngrams
- LSA/SVD

Feature Engineering

- TextLength, etc
- CosineSimilarity 
- count of: punctuations, ascii chars, USERNAMEs, EMOJIs, URLs
- tweet starts with USERNAME, or EMOJI, or URL, etc.
- count of swear words, or negative words, or positive words, etc. (need lists)
- semantic analysis research

Feature Selection

- Random Forest + VarImp plot 
- LASSO

Modeling

- SVM/SVC
- SGD
- Decision Trees


Direction

- define goal and purpose.. WHAT'S THE USE CASE?
- define evaluation metric better?
- if the goal is overall performance, the MCC (from sklearn.metrics import matthews_corrcoef) is probably preferred as it takes into account the entirety of the confusion matrix

---
