==> sentiment140/04_Twitter_Sentiment_Analysis_PREPROCESS.ipynb <==
## Twitter Sentiment Analysis 

---

### Pre-process cleaned data for machine learning 

While cleanup involved simply reformatting a Tweet's text by standardizing it and reducing the feature space (less punctuation, replacing usernames and URLs, lower casing, tokenizing, lemmatizing, etc.), pre-processing for machine learning is often more involved. It mainly consists of further data cleanup steps such as imputing NAs, but also feature engineering, and perhaps most importantly, a method of representing text in numerical form, such as [Document Term Matrices](./01_Document_Term_Matrices.ipynb), since most machine-learning algorithms do not accept text as input.

---

### Load cleaned TRAIN data


import os 
import re
import time

import numpy as np
import pandas as pd
import scipy.sparse as sp

# for ML preprocessing
from sklearn import preprocessing
from sklearn.feature_extraction.text import TfidfVectorizer
from sklearn.feature_extraction.text import CountVectorizer

# custom (see script)
import loading_module as lm

start_time = time.time()
df = lm.load_clean_data('X_train')
mins, secs = divmod(time.time() - start_time, 60)
print(f'Elapsed Time: {mins:0.0f} minute(s) and {secs:0.0f} second(s)')

### Quick EDA

df.shape

df.head() 

# load original train indices and subset
raw_path = os.path.join("..","data","1_raw","sentiment140")  
train_ix = np.load(os.path.join(raw_path, "train_ix.npy"))
df.index = list(train_ix)

df.head()

Look for emojis, get `NaNs` instead:

error_ix = []
emoji_ix = []
for i, tweet in enumerate(df['lemmatized'][:len(df)]):
    try:
        m = re.search(r'EMOJI', tweet)
        if m:
            emoji_ix.append(i)
    except TypeError as e: 
        error_ix.append(i)

# only a few NaNs
df.iloc[error_ix, ]

# emojis
df.iloc[emoji_ix[:5], ]

len(emoji_ix) # could be better?

### Impute NAs created during cleanup

We do not want to drop since the fact they ended up as empty strings is possibly informative.

df.isnull().sum()

# Impute with NULL as a string?
error_ix = []
NULL_ix = []
for i, tweet in enumerate(df['lemmatized'][:len(df)]):
    try:
        m = re.search(r'NULL', tweet)
        if m:
            NULL_ix.append(i)
    except TypeError as e: 
        error_ix.append(i)

df.iloc[NULL_ix, ]

NA_ix = df.loc[df['lemmatized'].isnull(), ].index
df['lemmatized'].loc[list(NA_ix), ] = 'NULL'

# double check
df.isnull().sum()

error_ix = []
NULL_ix = []
for i, tweet in enumerate(df['lemmatized'][:len(df)]):
    try:
        m = re.search(r'NULL', tweet)
        if m:
            NULL_ix.append(i)
    except TypeError as e: 
        error_ix.append(i)

df.iloc[NULL_ix, ]

### Create BoW DFM

# target
# load target...... why?

#y = np.array(df.iloc[:, 0]).ravel()

# lemmatized column (as array)
X_array = np.array(df.iloc[:, 2]).ravel()

X_array

# # create a BoW DFM 
bow_vectorizer = CountVectorizer()

start_bow = time.time()

X_bow = bow_vectorizer.fit_transform(X_array)

mins, secs = divmod(time.time() - start_bow, 60)
print(f"BoW vectorization time: {mins:0.0f} minute(s) and {secs:0.0f} second(s).")

X_bow

# visualize tiny portion, sparse indeed
X_bow[0:10, 0:20].todense()

# use log(tf)? do this now? what about savind idfs for test set?
# tfidf_vectorizer = TfidfVectorizer(sublinear_tf=True) 

### Save pre-processed DFMs

savepath = os.path.join("..","data","3_processed","sentiment140")
filename = 'X_bow'
filepath = os.path.join(savepath, ''.join([filename, '.npz']))
filepath

sp.save_npz(filepath, X_bow)

# save y target vector == we already have this!?
#np.save(os.path.join(dirpath, 'y'), y)

---
