==> sentiment140/06_Twitter_Sentiment_Analysis_LogisticRegression.ipynb <==
## Twitter Setiment Analysis 

### Part 4: Train and Evaluate Selected Models


First - take a training/modeling pause, assess situation:
 
TODO: feature extraction (text length?)

TODO: feature selection?

TODO: N-grams?

### Load Data


import os 
import time
import functools
import numpy as np
import scipy.sparse as sp
from joblib import dump, load

def timer(func):
    @functools.wraps(func)
    def wrapper_timer(*args, **kwargs):
        tic = time.perf_counter()
        value = func(*args, **kwargs)
        toc = time.perf_counter()
        elapsed_time = toc - tic
        print(f"Elapsed time: {elapsed_time:0.4f} seconds")
        return value
    return wrapper_timer

dirpath = os.path.join("..","data","3_processed","sentiment140")
filenames = ['X_text_bow'
            ,'X_tokn_bow'
            ,'X_filt_bow'
            ,'X_stem_bow'
            ,'X_text_tfidf'
            ,'X_tokn_tfidf'
            ,'X_filt_tfidf'
            ,'X_stem_tfidf'
            ,'X_text_log_tfidf'
            ,'X_tokn_log_tfidf'
            ,'X_filt_log_tfidf'
            ,'X_stem_log_tfidf']

filepaths = [os.path.join(dirpath, ''.join([filename, '.npz'])) for filename in filenames]

# load pre-processed featre vectors
@timer
def load_data():
    X_text_bow       = sp.load_npz(filepaths[0])
    X_tokn_bow       = sp.load_npz(filepaths[1])
    X_filt_bow       = sp.load_npz(filepaths[2])
    X_stem_bow       = sp.load_npz(filepaths[3])
    X_text_tfidf     = sp.load_npz(filepaths[4])
    X_tokn_tfidf     = sp.load_npz(filepaths[5])
    X_filt_tfidf     = sp.load_npz(filepaths[6])
    X_stem_tfidf     = sp.load_npz(filepaths[7])
    X_text_log_tfidf = sp.load_npz(filepaths[8])
    X_tokn_log_tfidf = sp.load_npz(filepaths[9])
    X_filt_log_tfidf = sp.load_npz(filepaths[10])
    X_stem_log_tfidf = sp.load_npz(filepaths[11])
    
    return (  X_text_bow      
            , X_tokn_bow      
            , X_filt_bow      
            , X_stem_bow      
            , X_text_tfidf    
            , X_tokn_tfidf    
            , X_filt_tfidf    
            , X_stem_tfidf    
            , X_text_log_tfidf
            , X_tokn_log_tfidf
            , X_filt_log_tfidf
            , X_stem_log_tfidf
            )

( X_text_bow      
, X_tokn_bow      
, X_filt_bow      
, X_stem_bow      
, X_text_tfidf    
, X_tokn_tfidf    
, X_filt_tfidf    
, X_stem_tfidf    
, X_text_log_tfidf
, X_tokn_log_tfidf
, X_filt_log_tfidf
, X_stem_log_tfidf
) = load_data()

# load y target vector
y = np.load(os.path.join(dirpath, 'y.npy'))

### Load Pre-Trained Models

dirpath = os.path.join("..","data","4_models","sentiment140")
os.listdir(dirpath)

filename = os.listdir(dirpath)[0]
filepath = os.path.join(dirpath, filename)
filepath

svm_clf = load(filepath)

REPRODUCIBILITY ISSUE: need to save exact train, test sets if using random sampling, or reproducible steps to get the same indices.

### Evaluate
