==> sentiment140/10.extra_GridSearchCV_DecisionTreeClassifier.ipynb <==
### Simple GridsearchCV with DecisionTreeClassifer 

End of chapter (6) exercise from Aurelien Geron's famous *Hands-On Machine Learning with Scikit-Learn & Tensorflow*.

from sklearn.datasets import make_moons

X, y = make_moons(n_samples=10000, noise=0.4, random_state=42)

X, y

from sklearn.model_selection import train_test_split

X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.2, random_state=42)

import numpy as np

# get notion of max depth's theoretical limit
print(f'Approx. tree depth without restrictions: {np.ceil(np.log2(len(X_train))):0.0f}')

from sklearn.model_selection import GridSearchCV
from sklearn.tree import DecisionTreeClassifier

params = {'max_leaf_nodes': list(range(2, 200)), 
          'min_samples_split': [2, 3, 4, 5, 6]}

grid_search_cv = GridSearchCV(DecisionTreeClassifier(random_state=42), 
                              params, n_jobs=6, verbose=1, cv=3)

grid_search_cv.fit(X_train, y_train)

grid_search_cv.best_estimator_

sorted(grid_search_cv.cv_results_.keys())

**A. Geron**: By default, `GridSearchCV` trains the best model found on the whole training set (you can change this by setting `refit=False`), so we don't need to do it again. We can simply evaluate the model's accuracy:

from sklearn.metrics import accuracy_score

y_pred = grid_search_cv.predict(X_test)
accuracy_score(y_test, y_pred)

for i,v in enumerate(grid_search_cv.cv_results_['mean_test_score']):
    if v == max(grid_search_cv.cv_results_['mean_test_score']):
        print('Max mean test accuracy:', round(v,4), \
              '\nParams:', grid_search_cv.cv_results_['params'][i])

# train with those params in entire training set
dtree = DecisionTreeClassifier(random_state=42, min_samples_split=2, # or 3, 4, 5, 6...
                               max_leaf_nodes=17)
dtree.fit(X_train, y_train)

# predict on test set
y_preds = dtree.predict(X_test)

# print accuracy on test target
accuracy_score(y_test, y_preds)

---

### Grow a Random Forest

With 1000 trees of 100 instances each.

from sklearn.model_selection import ShuffleSplit

subsets = []

rs = ShuffleSplit(n_splits=1000, test_size=len(X_train) - 100, random_state=42)
for train_sub_ix, test_sub_ix in rs.split(X_train):
    X_sub_train = X_train[train_sub_ix]
    y_sub_train = y_train[train_sub_ix]
    subsets.append((X_sub_train, y_sub_train))

subsets[0][0][:10], subsets[0][1][:10]

from sklearn.base import clone

forest = [clone(grid_search_cv.best_estimator_) for _ in range(1000)]

forest[:5]

accuracy_scores = []

# train 1000 decision tree classifiers (but... are't they all the same? Why do we need to clone...)
for tree, (X_sub_train, y_sub_train) in zip(forest, subsets):
    tree.fit(X_sub_train, y_sub_train)
    
    y_pred = tree.predict(X_test)
    accuracy_scores.append(accuracy_score(y_test, y_pred))

np.mean(accuracy_scores)

# my old code
X_rf, y_rf = [], []
for i,v in enumerate(rs.split(X_train, y_train)):
    X_rf.append(X_train[v[0]])
    y_rf.append(y_train[v[0]])

clf = DecisionTreeClassifier(random_state=42, max_leaf_nodes=17)

accs = []
for ix, tree in enumerate(X_rf):
    clf.fit(tree, y_rf[ix])
    y_pred = clf.predict(X_test)
    accs.append(accuracy_score(y_test, y_pred))

np.mean(accs)

**Magic:** 

- for each test set instance, generate the predictions of the 1000 trees 
- keep only the most frequent prediction (the *mode*)

This procedure gives you the majority-vote predictions over the test set.

Y_pred = np.empty([1000, len(X_test)], dtype=np.uint8)

Y_pred.shape # just a 1k by 2k empty matrix

for tree_ix, tree in enumerate(forest):
    Y_pred[tree_ix] = tree.predict(X_test)

forest[0].predict(X_test)

forest[1].predict(X_test)

from scipy.stats import mode

y_pred_majority_votes, n_votes = mode(Y_pred, axis=0)

accuracy_score(y_test, y_pred_majority_votes.reshape([-1]))

# my old code
from scipy import stats

y_preds = []
for ix, tree in enumerate(X_rf):
    clf.fit(tree, y_rf[ix])
    y_preds.append(clf.predict(X_test))

y_preds_array = np.vstack(y_preds)

majority_vote = stats.mode(y_preds_array)

majority_vote[0][0]

accuracy_score(y_test, majority_vote[0][0])

---
